Added skeleton in viewModel: the function and var i would need. reviews will be the reviews that are fetched from json file and stored as CoffeeShop. I have added a function fetchReviews() to do the json parsing. I prefer the functions to return a value and save it where and when the function is called. I have also navigated protocols and struct to be a part of the view model. functions and variables will be added to viewModel as and when needed. currently i need these 2 for now.
