+++ interview-project-master/SNInterview/CoffeeShopViewModel.swift
@@ -7,3 +7,28 @@
 //
 
 import Foundation
+import UIKit
+
+struct CoffeeShop {
+    let name: String
+    let review: String
+    let rating: Int
+}
+
+protocol CoffeeShopTapDelegate {
+    func didSelectItem(_ item: UIView?)
+}
+
+class CoffeeShopViewModel: NSObject {
+    var reviews: [CoffeeShop]!
+    
+    override init() {
+        super.init()
+        reviews = fetchReviews()
+    }
+    
+    func fetchReviews() -> [CoffeeShop] {
+        var fetchedReviews = [CoffeeShop]()
+        return fetchedReviews
+    }
+}
